// projects/codx-ac/src/lib/models/PaymentOrderLines.model.ts
export class PaymentOrderLines{
    recID: any;
	transID: any;
	rowNo: any;
	reasonID: any;
	accountID: any;
	subLGType: any;
	objectType: any;
	objectID: any;
	settlement: any;
	settledID: any;
	settledNo: any;
	contractID: any;
	dr: any;
	cr: any;
	dr2: any;
	cr2: any;
	diM1: any;
	diM2: any;
	diM3: any;
	projectID: any;
	offsetAcctID: any;
	isBrigdeAcct: any;
	vatID: any;
	vatAmt: any;
	vatAmt2: any;
	illegalAmt: any;
	assetType: any;
	assetGroupID: any;
	serviceDate: any;
	servicePeriods: any;
	employeeID: any;
	siteID: any;
	refType: any;
	refNo: any;
	refLineID: any;
	note: any;
	singleEntry: any;
	autoCreated: any;
	createdOn: any;
	createdBy: any;
	modifiedOn: any;
	modifiedBy: any;
}